DAOS-8921 Coverity: Various fixes for Coverity Issues
Coverity identified multiple issues in various C sources. This fixes CIDs:

329561, 331438, 332668, 332945, 339129, 340852

+++ src/tests/suite/daos_rebuild_ec.c
@@ -898,7 +898,7 @@ rebuild_ec_snapshot(void **state, daos_oclass_id_t oclass, int shard)
 	daos_pool_set_prop(arg->pool.pool_uuid, "reclaim", "time");
 	oid = daos_test_oid_gen(arg->coh, oclass, 0, 0, arg->myrank);
 	ioreq_init(&req, arg->coh, oid, DAOS_IOD_ARRAY, arg);
-	data_size = ec_data_nr_get(oid) * EC_CELL_SIZE + 1000;
+	data_size = ec_data_nr_get(oid) * (uint64_t)EC_CELL_SIZE + 1000;
 	data = (char *)malloc(data_size);
 	assert_true(data != NULL);
 	verify_data = (char *)malloc(data_size);
